requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.1
django-ckeditor-5==0.2.13
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-js-asset==2.2.0
django-mptt==0.16.0
django-multiupload==0.6.1
django-phonenumber-field==8.0.0
django-reversion==5.1.0
djangorestframework==3.15.2
idna==3.8
magic-filter==1.0.12
multidict==6.0.5
netaddr==1.3.0
phonenumbers==8.13.43
pilkit==3.0
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
rollbar==1.0.0
six==1.16.0
sorl-thumbnail==12.10.0
soupsieve==2.6
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
wrapt==1.16.0
yarl==1.9.4
